Moving handlebars output to separate file and function.
