+++ templates/openacr-markdown-0.1.0.handlebars
@@ -104,7 +104,7 @@ Notes: {{data-category.notes}}
 {{#if related_openacrs}}
 ## Related OpenACRs
 {{#each related_openacrs as |related-openacr|}}
-* {{related-openacr.url}} ({{related-openacr.type}})
+- {{related-openacr.url}} ({{related-openacr.type}})
 {{/each}}
 {{/if}}
