+++ tests/examples/valid.html
@@ -27,7 +27,7 @@
         <a href="#report-date" aria-hidden="true" class="header-anchor">#</a>
         Report Date
       </h2>
-        11/16/2021
+        11/23/2021
 
         <h2 id="product-description">
           <a href="#product-description" aria-hidden="true" class="header-anchor">#</a>
