+++ tests/examples/valid.markdown
@@ -6,7 +6,7 @@ Based on VPAT® 2.4 Revised Section 508 Edition
 Lorem Ipsum 1.1
 
 ## Report Date
-11/16/2021
+11/23/2021
 
 ## Product Description
 Fake text
